+++ nflow-engine/src/test/java/io/nflow/engine/internal/executor/WorkflowLifecycleTest.java
@@ -1,4 +1,4 @@
-package io.nflow.engine.internal.config;
+package io.nflow.engine.internal.executor;
 
 import static java.lang.Boolean.TRUE;
 import static org.hamcrest.Matchers.is;
@@ -17,6 +17,7 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.core.env.Environment;
 
 import io.nflow.engine.internal.executor.WorkflowDispatcher;
+import io.nflow.engine.internal.executor.WorkflowLifecycle;
 
 @RunWith(MockitoJUnitRunner.class)
 public class WorkflowLifecycleTest {
